Move top-level `CoinSelection` module to `Cardano.Tx.Balance.Internal`.
